% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airline_dt.R
\name{airline_dt}
\alias{airline_dt}
\title{Airline Safety Rating DataTable}
\usage{
airline_dt(input_rate)
}
\arguments{
\item{input_rate}{Shiny user-interface input}
}
\value{
A DataTable with 2 columns : Airline Name and Tail Logo
}
\description{
Create interactive Data Table of Airlines based on the rating input by the user.
}
