% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airline_pic.R
\name{airline_pic}
\alias{airline_pic}
\title{Airline Tail Logo Tag}
\usage{
airline_pic(input)
}
\arguments{
\item{input}{Shiny user-interface input}
}
\description{
Create a HTML IMG tag for the tail logo of the airlines for output in Shiny APP.
}
