% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airline_plot.R
\name{airline_plot}
\alias{airline_plot}
\title{Airline Incidents Bar Plot}
\usage{
airline_plot(input_aero, input_event)
}
\arguments{
\item{input_aero}{Shiny user-interface input}

\item{input_event}{Shiny user-interface input}
}
\description{
Create interactive bar plot comparing number of accidents in two different
range of year of an airline based on the input of the selected type of events.
}
