% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airline_rating.R
\name{airline_rating}
\alias{airline_rating}
\title{State Airline Safety Rating}
\usage{
airline_rating(name)
}
\arguments{
\item{name}{an airline name listed in the safetydata, typed in both lowercase or uppercase is acceptable}
}
\description{
State the rating an airline out of the 56 listed in the safetydata based on the input.
}
\examples{
airline_rating("All Nippon Airways")
airline_rating("QANTAS")
airline_rating("garuda indonesia")

}
