% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airline_rating0.R
\name{airline_rating0}
\alias{airline_rating0}
\title{State Airline Safety Rating}
\usage{
airline_rating0(name = NULL)
}
\arguments{
\item{name}{an airline name listed in the safetydata, typed in both lowercase or uppercase, even mixed, is acceptable}
}
\description{
Calculate the rating an airline out of the 56 listed in the safetydata based on the input.
}
\examples{
airline_rating("All Nippon Airways")
airline_rating("QANTAS")
airline_rating("garuda inDOnesia")

}
