% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsafety-package.R
\docType{package}
\name{airsafety-package}
\alias{airsafety}
\alias{airsafety-package}
\title{airsafety: What the Package Does (One Line, Title Case)}
\description{
Airsafety package contains the airline safety data and includes a Shiny Application for user to interactively learn more about safety of airlines around the world.
}
\author{
\strong{Maintainer}: Josephine Valensia \email{jval0011@student.monash.edu}

}
\keyword{internal}
