% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetydata_ratePic.R
\name{safetydata_ratePic}
\alias{safetydata_ratePic}
\title{Safetydata with Ratings and Logo}
\usage{
safetydata_ratePic(data)
}
\arguments{
\item{data}{airsafety::safetydata dataset}
}
\description{
Transform the airesafety::safetydata dataset into wide-format and add new
columns to calculate rating and add urls for airline tail logo, creating
a modified dataset with 56 rows and 14 variables.
}
